# [ArrowStringArray] PERF: isin using native pyarrow function if available