Unlike MaskedArray, this returns a numpy bool array to be consistent with the EA interface and StringArray and also due to the fact that the returned boolean array has no null values to be consistent with the latest version of pyarrow.


```
[  0.00%] ·· Benchmarking existing-py_home_simon_miniconda3_envs_pandas-dev_bin_python
[  4.17%] ··· algos.isin.IsIn.time_isin                                                                                                                   ok
[  4.17%] ··· ================== ==========
                    dtype                  
              ------------------ ----------
                    int64         295±0μs  
                    uint64        348±0μs  
                    object        337±0μs  
                    Int64         785±0μs  
                   boolean        868±0μs  
                     bool         420±0μs  
                datetime64[ns]    4.67±0ms 
               category[object]   9.46±0ms 
                category[int]     7.30±0ms 
                     str          535±0μs  
                    string        556±0μs  
                 arrow_string     330±0μs  
              ================== ==========

[  8.33%] ··· algos.isin.IsIn.time_isin_categorical                                                                                                       ok
[  8.33%] ··· ================== ==========
                    dtype                  
              ------------------ ----------
                    int64         374±0μs  
                    uint64        507±0μs  
                    object        467±0μs  
                    Int64         633±0μs  
                   boolean        702±0μs  
                     bool         458±0μs  
                datetime64[ns]    3.10±0ms 
               category[object]   10.2±0ms 
                category[int]     9.12±0ms 
                     str          598±0μs  
                    string        628±0μs  
                 arrow_string     404±0μs  
              ================== ==========

[ 12.50%] ··· algos.isin.IsIn.time_isin_empty                                                                                                             ok
[ 12.50%] ··· ================== ==========
                    dtype                  
              ------------------ ----------
                    int64         275±0μs  
                    uint64        285±0μs  
                    object        390±0μs  
                    Int64         1.06±0ms 
                   boolean        1.14±0ms 
                     bool         280±0μs  
                datetime64[ns]    295±0μs  
               category[object]   4.17±0ms 
                category[int]     3.49±0ms 
                     str          424±0μs  
                    string        718±0μs  
                 arrow_string     140±0μs  
              ================== ==========

[ 16.67%] ··· algos.isin.IsIn.time_isin_mismatched_dtype                                                                                                  ok
[ 16.67%] ··· ================== ==========
                    dtype                  
              ------------------ ----------
                    int64         221±0μs  
                    uint64        216±0μs  
                    object        337±0μs  
                    Int64         337±0μs  
                   boolean        356±0μs  
                     bool         323±0μs  
                datetime64[ns]    348±0μs  
               category[object]   4.86±0ms 
                category[int]     3.47±0ms 
                     str          586±0μs  
                    string        525±0μs  
                 arrow_string     224±0μs  
              ================== ==========
```